+++ alpsqutip/meanfield/__init__.py
@@ -1,9 +1,4 @@
 from .meanfield import project_meanfield
-from .projections import (
-    n_body_projection,
-    one_body_from_qutip_operator,
-    project_operator_to_m_body,
-)
 from .self_consistent_projections import self_consistent_project_meanfield
 from .variational import variational_quadratic_mfa
